+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `ref_range_separator` option to configure the character used for reference
     ranges
   - `use_ref_ranges` alias for `use_alt`
+  - New *kicad* format to mimic KiCad's internal BoM.
 - PCB Print: a mechanism to filter components for a particular layer (#706)
 
 ### Fixed